requirements.txt
fastapi
uvicorn
streamlit
openai
python-dotenv
pydantic
requests